vita: update for latest changes, needs testing

+++ source/platforms/sdl2/sdl2_audio.cpp
@@ -80,9 +80,9 @@ void SDL2Audio::play(const void *data, int len, bool sync) {
             }
         } else {
             // Clear the audio queue arbitrarily to avoid it backing up too far
-            if (SDL_GetQueuedAudioSize(deviceID) > (Uint32) (len * 5)) {
+            if (SDL_GetQueuedAudioSize(deviceID) > (size_t) (len * 5)) {
                 printf("SDL2Audio::play: queue is full (buffer len = %i, queued = %i)\n",
-                       len, SDL_GetQueuedAudioSize(deviceID));
+                       len, (size_t) SDL_GetQueuedAudioSize(deviceID));
                 SDL_ClearQueuedAudio(deviceID);
             }
         }
